requirements.txt
aiosmtplib==1.1.6
alabaster==0.7.13
alembic==1.13.0
aniso8601==9.0.1
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attr==0.3.2
attrs==23.1.0
Babel==2.13.1
blinker==1.7.0
boltons==23.1.1
boto3==1.29.2
botocore==1.32.2
Brotli==1.1.0
brotlicffi==1.1.0.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
configparser==6.0.0
contextlib2==21.6.0
cryptography==41.0.5
Cython==3.0.6
decorator==5.1.1
Deprecated==1.2.13
dl==0.1.0
docutils==0.20.1
exceptiongroup==1.1.3
filelock==3.13.1
Flask==2.2.5
Flask-Cors==3.0.10
Flask-Migrate==4.0.5
flask-restx==1.2.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
funcsigs==1.0.2
h11==0.14.0
healthcheck==1.3.3
HTMLParser==0.0.2
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
logging==0.4.9.6
Mako==1.3.0
MarkupSafe==2.1.3
mock==5.1.0
packaging==23.2
phonenumbers==8.12.48
Pillow==10.1.0
pkce==1.0.3
protobuf==4.25.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.10.4
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyrsistent==0.20.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
railroad==0.5.0
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
s3transfer==0.7.0
schedule==0.6.0
sentry-sdk==1.35.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.23
supertokens-python==0.17.0
tldextract==3.1.0
toolz==0.12.0
trove-classifiers==2023.11.29
twilio==7.9.1
typing==3.7.4.3
typing_extensions==4.8.0
urllib3==1.26.18
validators==0.20.0
warn==0.1.0
Werkzeug==2.3.8
wrapt==1.16.0
zipp==3.17.0

# fmlib requirements
-r ./fmlib/requirements.txt
